refactor(Power): Power.java -> MovementStrategy.java 로 변경
