+++ src/main/java/racingcar/RacingCarApplication.java
@@ -3,7 +3,7 @@ package racingcar;
 import racingcar.domain.game.GameInfo;
 import racingcar.domain.game.GameResult;
 import racingcar.domain.game.RacingGame;
-import racingcar.domain.car.power.RandomPower;
+import racingcar.domain.car.strategy.RandomMovementStrategy;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
@@ -16,7 +16,7 @@ public class RacingCarApplication {
 
         RacingGame racingGame = new RacingGame(gameInfo);
 
-        racingGame.startGame(new RandomPower());
+        racingGame.startGame(new RandomMovementStrategy());
 
         GameResult gameResult = racingGame.getGameResult();
