--- src/main/java/racingcar/domain/car/power/Power.java
@@ -1,6 +1,0 @@
-package racingcar.domain.car.power;
-
-@FunctionalInterface
-public interface Power {
-    boolean isMovable();
-}
