+++ src/main/java/racingcar/domain/car/strategy/MovementStrategy.java
@@ -1,0 +1,6 @@
+package racingcar.domain.car.strategy;
+
+@FunctionalInterface
+public interface MovementStrategy {
+    boolean isMovable();
+}
