+++ src/test/java/racingcar/domain/car/ParticipateCarsTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import racingcar.domain.car.power.FixedPower;
+import racingcar.domain.car.strategy.FixedMovementStrategy;
 
 import java.util.Arrays;
 import java.util.List;
@@ -27,7 +27,7 @@ public class ParticipateCarsTest {
     @MethodSource("provideParticipateCarsInformation")
     void tryMove(int numberOfCar, List<Integer> moveResult) {
         ParticipateCars participateCars = new ParticipateCars(numberOfCar);
-        List<Integer> actualResult = participateCars.moveCars(new FixedPower());
+        List<Integer> actualResult = participateCars.moveCars(new FixedMovementStrategy());
 
         assertThat(actualResult).isEqualTo(moveResult);
